abstract dir traversal in prep for testing
