+++ src/find-plus.mjs
@@ -1,9 +1,9 @@
-import * as fs from 'node:fs/promises'
 import * as fsPath from 'node:path'
 
 import { addImpliedTests } from './lib/add-implied-tests'
 import { checkRoot } from './lib/check-root'
 import { alphaSorter, breadthFirstSorter, depthFirstSorter } from './lib/sorters'
+import { traverseDirs } from './lib/traverse-dirs'
 import { verifyParams } from './lib/verify-params'
 
 const find = async(params = {}) => {
@@ -16,9 +16,6 @@ const find = async(params = {}) => {
   }
 
   const {
-    depth,
-    excludeRoot = false,
-    noTraverseFailed = false,
     root = throw new Error("Must provide 'root' to find."),
     sort,
     tests = []
@@ -35,53 +32,8 @@ const find = async(params = {}) => {
   rootStat.name = fsPath.basename(root)
   rootStat.depth = 0
 
-  const accumulator = []
-  let currDepth = 0
-
-  let frontier = []
-  if (excludeRoot === true && noTraverseFailed === false) { // no need for tests
-    frontier.push(rootStat)
-  }
-  else {
-    const rootPasses = !myTests.some((t) => !t(rootStat, currDepth))
-    if (excludeRoot === false && rootPasses) {
-      accumulator.push(rootStat)
-    }
-    if (rootPasses || noTraverseFailed === false) {
-      frontier.push(rootStat)
-    }
-  }
-
-  currDepth += 1
-  // eslint-disable-next-line no-unmodified-loop-condition
-  while ((depth === undefined || depth >= currDepth) && frontier.length > 0) {
-    const newFrontier = []
-    for (const dirEnt of frontier) {
-      const dirPath = fsPath.join(dirEnt.parentPath, dirEnt.name)
-      const files = await fs.readdir(dirPath, { withFileTypes : true })
-      for (const file of files) {
-        file.depth = currDepth
-
-        // node 19.x DirEnt's lack parent path
-        if (file.parentPath === undefined) {
-          file.parentPath = dirPath
-        }
-
-        const pass = !myTests.some((t) => !t(file, currDepth))
-
-        if (file.isDirectory() && (pass || noTraverseFailed === false)) {
-          console.log('dirPath:', dirPath, 'adding frontier:', file) // DEBUG
-          newFrontier.push(file)
-        }
-        if (pass) {
-          accumulator.push(file)
-        }
-      }
-    }
-    frontier = newFrontier
-
-    currDepth += 1
-  }
+  // params need to come first, we override root and tests
+  const matchedFiles = await traverseDirs({ ...params, root : rootStat, tests : myTests })
 
   // results in depth-first sort of full directory paths
   if (sort !== 'none') {
@@ -90,10 +42,10 @@ const find = async(params = {}) => {
       : sort === 'alpha'
         ? alphaSorter
         : /* default */ breadthFirstSorter
-    accumulator.sort(sorter)
+    matchedFiles.sort(sorter)
   }
 
-  const result = accumulator.map(({ name, parentPath }) => fsPath.resolve(parentPath, name))
+  const result = matchedFiles.map(({ name, parentPath }) => fsPath.resolve(parentPath, name))
 
   return result
 }
